Prove that Asm_coop/eff is deterministic (except for one whole, where the semantics is underconstrained).
The problem is with the use of external_call. This function
(from CompCert) takes a trace argument which we no longer need.
The result is that the following lemma doesn't hold:

Lemma external_call_trace_eq ef args m0 t v m t' v' m' :
  external_call ef ge args m0 t v m ->
  external_call ef ge args m0 t' v' m' ->
  t=t'.

The fix: use a modified version of external_call that
does not specify the trace of a nonobservable external.